+++ application/wypl-core/src/main/java/com/wypl/wyplcore/review/service/ReviewReadService.java
@@ -7,10 +7,10 @@ import com.wypl.wyplcore.review.data.response.ReviewDetailResponse;
 import com.wypl.wyplcore.review.data.response.ReviewListResponse;
 
 public interface ReviewReadService {
-	ReviewDetailResponse getDetailReview(int memberId, int reviewId);
+	ReviewDetailResponse getDetailReview(long memberId, long reviewId);
 
-	ReviewListResponse getReviews(int memberId, Integer lastReviewId, ReviewType reviewType, LocalDate startDate,
+	ReviewListResponse getReviews(long memberId, Long lastReviewId, ReviewType reviewType, LocalDate startDate,
 		LocalDate endDate);
 
-	ReviewListResponse getReviewsByScheduleId(int memberId, int scheduleId, ReviewType reviewType);
+	ReviewListResponse getReviewsByScheduleId(long memberId, long scheduleId, ReviewType reviewType);
 }
